Move custom slider component API to the component ts file
