+++ renderers/angular/a2ui_explorer/src/app/demo-catalog.ts
@@ -17,18 +17,8 @@
 import { Injectable } from '@angular/core';
 import { z } from 'zod';
 import { BasicCatalogBase, BASIC_FUNCTIONS } from '@a2ui/angular/v0_9';
-import { CustomSliderComponent } from './custom-slider.component';
-import { ComponentApi, createFunctionImplementation, FunctionImplementation } from '@a2ui/web_core/v0_9';
-
-export const customSliderApi = {
-  name: 'CustomSlider',
-  schema: z.object({
-    label: z.string().optional(),
-    value: z.number().optional(),
-    min: z.number().optional(),
-    max: z.number().optional(),
-  }) as any,
-} satisfies ComponentApi;
+import { CustomSliderComponentDeclaration } from './custom-slider.component';
+import { createFunctionImplementation, FunctionImplementation } from '@a2ui/web_core/v0_9';
 
 /**
  * A catalog specific to the demo, extending the basic catalog with custom components.
@@ -57,7 +47,7 @@ export class DemoCatalog extends BasicCatalogBase {
     super({
       id: 'https://a2ui.org/specification/v0_9/basic_catalog.json',
       components: {},
-      extraComponents: [{ ...customSliderApi, component: CustomSliderComponent }],
+      extraComponents: [CustomSliderComponentDeclaration],
       functions,
     });
   }
